agents/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.2
fastapi==0.114.2
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
modal==0.64.110
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.1.0
packaging==24.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
protobuf==4.25.4
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
rapidfuzz==3.9.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.1
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
starlette==0.38.5
synchronicity==0.7.6
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
together==1.2.12
trove-classifiers==2024.5.17
typer==0.12.5
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
urllib3==2.2.1
virtualenv==20.26.2
watchfiles==0.24.0
xattr==1.1.0
yarl==1.11.1
zipp==3.18.2
